初始化v1.2版本

+++ jeesns-cms/jeesns-cms-web/src/main/java/com/lxinet/jeesns/cms/web/manage/ArticleController.java
@@ -1,57 +1,43 @@
 package com.lxinet.jeesns.cms.web.manage;
 
-import com.lxinet.jeesns.commons.utils.MemberUtil;
-import com.lxinet.jeesns.member.interceptor.AdminLoginInterceptor;
-import com.lxinet.jeesns.core.annotation.Before;
+import com.lxinet.jeesns.core.model.MemberToken;
 import com.lxinet.jeesns.core.dto.ResponseModel;
 import com.lxinet.jeesns.core.model.Page;
 import com.lxinet.jeesns.core.web.BaseController;
-import com.lxinet.jeesns.cms.model.ArticleCate;
 import com.lxinet.jeesns.cms.model.Article;
-import com.lxinet.jeesns.cms.service.IArticleCateService;
 import com.lxinet.jeesns.cms.service.IArticleService;
 import com.lxinet.jeesns.member.model.Member;
+import com.lxinet.jeesns.member.service.IMemberService;
+import org.apache.ibatis.annotations.Param;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
-
 import javax.annotation.Resource;
 import javax.validation.Valid;
-import java.util.List;
 
 /**
  * Created by zchuanzhao on 16/9/29.
  */
 @Controller("manageArticleController")
 @RequestMapping("/")
-@Before(AdminLoginInterceptor.class)
 public class ArticleController extends BaseController {
-    private static final String MANAGE_FTL_PATH = "/manage/cms/article/";
-    @Resource
-    private IArticleCateService articleCateService;
     @Resource
     private IArticleService articleService;
+    @Resource
+    private IMemberService memberService;
 
-    @RequestMapping("${managePath}/cms/index")
-    public String index(String key, @RequestParam(value = "cateid",defaultValue = "0",required = false) Integer cateid,
-    @RequestParam(value = "status",defaultValue = "2",required = false) Integer status,
-    @RequestParam(value = "memberId",defaultValue = "0",required = false) Integer memberId, Model model) {
-        List<ArticleCate> cateList = articleCateService.list();
+    @RequestMapping("${managePath}/cms/article/list")
+    @ResponseBody
+    public Object list(String key, @RequestParam(value = "cateId",defaultValue = "0",required = false) Integer cateId,
+                @RequestParam(value = "status",defaultValue = "2",required = false) Integer status,
+                @RequestParam(value = "memberId",defaultValue = "0",required = false) Integer memberId) {
+        ResponseModel<MemberToken> validMemberTokenModel = validMemberToken();
+        if (validMemberTokenModel.getCode() == -1){
+            return validMemberTokenModel;
+        }
         Page page = new Page(request);
-        ResponseModel responseModel = articleService.listByPage(page,key,cateid,status,memberId);
-        model.addAttribute("model",responseModel);
-        model.addAttribute("cateList",cateList);
-        model.addAttribute("key",key);
-        model.addAttribute("cateid",cateid);
-        return MANAGE_FTL_PATH + "index";
-    }
-
-    @RequestMapping(value="${managePath}/cms/article/add",method = RequestMethod.GET)
-    public String add(Model model) {
-        List<ArticleCate> cateList = articleCateService.list();
-        model.addAttribute("cateList",cateList);
-        return MANAGE_FTL_PATH + "add";
+        ResponseModel responseModel = articleService.listByPage(page,key,cateId,status,memberId);
+        return responseModel;
     }
 
     @RequestMapping(value="${managePath}/cms/article/save",method = RequestMethod.POST)
@@ -60,7 +46,11 @@ public class ArticleController extends BaseController {
         if(bindingResult.hasErrors()){
             return new ResponseModel(-1,getErrorMessages(bindingResult));
         }
-        Member loginMember = MemberUtil.getLoginMember(request);
+        ResponseModel<MemberToken> validMemberTokenModel = validMemberToken();
+        if (validMemberTokenModel.getCode() == -1){
+            return validMemberTokenModel;
+        }
+        Member loginMember = memberService.findById(validMemberTokenModel.getData().getMemberId());
         ResponseModel responseModel = articleService.save(loginMember,article);
         if(responseModel.getCode() == 0){
             responseModel.setCode(3);
@@ -68,36 +58,32 @@ public class ArticleController extends BaseController {
         return responseModel;
     }
 
-    @RequestMapping(value="${managePath}/cms/article/list",method = RequestMethod.GET)
-    public String list(String key, @RequestParam(value = "cateid",defaultValue = "0",required = false) Integer cateid,
-                       @RequestParam(value = "status",defaultValue = "2",required = false) Integer status,
-                       @RequestParam(value = "memberId",defaultValue = "0",required = false) Integer memberId, Model model) {
-        Page page = new Page(request);
-        ResponseModel responseModel = articleService.listByPage(page,key,cateid,status,memberId);
-        model.addAttribute("model",responseModel);
-        return MANAGE_FTL_PATH + "list";
-    }
-
-    @RequestMapping(value="${managePath}/cms/article/edit/{id}",method = RequestMethod.GET)
-    public String edit(@PathVariable("id") Integer id, Model model){
-        Member loginMember = MemberUtil.getLoginMember(request);
-        List<ArticleCate> cateList = articleCateService.list();
-        model.addAttribute("cateList",cateList);
+    @RequestMapping(value="${managePath}/cms/article/info",method = RequestMethod.GET)
+    @ResponseBody
+    public Object info(@Param("id") Integer id){
+        ResponseModel<MemberToken> validMemberTokenModel = validMemberToken();
+        if (validMemberTokenModel.getCode() == -1){
+            return validMemberTokenModel;
+        }
+        Member loginMember = memberService.findById(validMemberTokenModel.getData().getMemberId());
         Article article = articleService.findById(id,loginMember);
-        model.addAttribute("article",article);
-        return MANAGE_FTL_PATH + "/edit";
+        return article;
     }
 
     @RequestMapping(value="${managePath}/cms/article/update",method = RequestMethod.POST)
     @ResponseBody
     public Object update(@Valid Article article,BindingResult bindingResult) {
+        ResponseModel<MemberToken> validMemberTokenModel = validMemberToken();
+        if (validMemberTokenModel.getCode() == -1){
+            return validMemberTokenModel;
+        }
         if(bindingResult.hasErrors()){
             new ResponseModel(-1,getErrorMessages(bindingResult));
         }
         if(article.getId() == null){
             return new ResponseModel(-2);
         }
-        Member loginMember = MemberUtil.getLoginMember(request);
+        Member loginMember = memberService.findById(validMemberTokenModel.getData().getMemberId());
         ResponseModel responseModel = articleService.update(loginMember,article);
         if(responseModel.getCode() == 0){
             responseModel.setCode(3);
@@ -106,17 +92,25 @@ public class ArticleController extends BaseController {
     }
 
 
-    @RequestMapping(value = "${managePath}/cms/article/delete/{id}",method = RequestMethod.GET)
+    @RequestMapping(value = "${managePath}/cms/article/delete",method = RequestMethod.GET)
     @ResponseBody
-    public Object delete(@PathVariable("id") Integer id){
-        Member loginMember = MemberUtil.getLoginMember(request);
+    public Object delete(@Param("id") Integer id){
+        ResponseModel<MemberToken> validMemberTokenModel = validMemberToken();
+        if (validMemberTokenModel.getCode() == -1){
+            return validMemberTokenModel;
+        }
+        Member loginMember = memberService.findById(validMemberTokenModel.getData().getMemberId());
         ResponseModel response = articleService.delete(loginMember,id);
         return response;
     }
 
-    @RequestMapping(value = "${managePath}/cms/article/audit/{id}",method = RequestMethod.GET)
+    @RequestMapping(value = "${managePath}/cms/article/audit",method = RequestMethod.GET)
     @ResponseBody
-    public Object audit(@PathVariable("id") Integer id){
+    public Object audit(@Param("id") Integer id){
+        ResponseModel<MemberToken> validMemberTokenModel = validMemberToken();
+        if (validMemberTokenModel.getCode() == -1){
+            return validMemberTokenModel;
+        }
         ResponseModel response = articleService.audit(id);
         return response;
     }
