+++ jeesns-member/jeesns-member-web/src/main/java/com/lxinet/jeesns/member/web/manage/MemberController.java
@@ -1,20 +1,16 @@
 package com.lxinet.jeesns.member.web.manage;
 
-import com.lxinet.jeesns.commons.utils.MemberUtil;
-import com.lxinet.jeesns.member.interceptor.AdminLoginInterceptor;
-import com.lxinet.jeesns.core.annotation.Before;
+import com.lxinet.jeesns.core.model.MemberToken;
 import com.lxinet.jeesns.core.dto.ResponseModel;
 import com.lxinet.jeesns.core.model.Page;
 import com.lxinet.jeesns.core.web.BaseController;
 import com.lxinet.jeesns.member.model.Member;
 import com.lxinet.jeesns.member.service.IMemberService;
+import org.apache.ibatis.annotations.Param;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
-
 import javax.annotation.Resource;
 
 /**
@@ -22,9 +18,7 @@ import javax.annotation.Resource;
  */
 @Controller("manageMemberController")
 @RequestMapping("/")
-@Before(AdminLoginInterceptor.class)
 public class MemberController extends BaseController {
-    private static final String MANAGE_FTL_PATH = "/manage/mem/";
     @Resource
     private IMemberService memberService;
 
@@ -32,45 +26,38 @@ public class MemberController extends BaseController {
     /**
      * 管理员列表
      * @param key
-     * @param model
      * @return
      */
-    @RequestMapping(value = "${managePath}/mem/manager/list",method = RequestMethod.GET)
-    public String managerList(String key,Model model) {
-        Member loginMember = MemberUtil.getLoginMember(request);
+    @RequestMapping(value = "${managePath}/member/manager/list",method = RequestMethod.GET)
+    @ResponseBody
+    public Object managerList(String key) {
+        ResponseModel<MemberToken> validMemberTokenModel = validMemberToken();
+        if (validMemberTokenModel.getCode() == -1){
+            return validMemberTokenModel;
+        }
+        Member loginMember = memberService.findById(validMemberTokenModel.getData().getMemberId());
         if(loginMember.getIsAdmin() == 1){
-            return errorModel(model, "没有权限");
+            return new ResponseModel(-1, "没有权限");
         }
         Page page = new Page(request);
         ResponseModel responseModel = memberService.managerList(page,key);
-        model.addAttribute("model",responseModel);
-        model.addAttribute("key",key);
-        return MANAGE_FTL_PATH + "manager/list";
+        return responseModel;
     }
 
-    /**
-     * 授权管理员
-     * @param model
-     * @return
-     */
-    @RequestMapping(value = "${managePath}/mem/manager/add",method = RequestMethod.GET)
-    public String managerAdd(Model model) {
-        Member loginMember = MemberUtil.getLoginMember(request);
-        if(loginMember.getIsAdmin() == 1){
-            return errorModel(model, "没有权限");
-        }
-        return MANAGE_FTL_PATH + "manager/add";
-    }
 
     /**
      * 授权管理员
      * @param name
      * @return
      */
-    @RequestMapping(value = "${managePath}/mem/manager/add",method = RequestMethod.POST)
+    @RequestMapping(value = "${managePath}/member/manager/save",method = RequestMethod.POST)
     @ResponseBody
-    public Object managerAdd(String name) {
-        Member loginMember = MemberUtil.getLoginMember(request);
+    public Object managerSave(String name) {
+        ResponseModel<MemberToken> validMemberTokenModel = validMemberToken();
+        if (validMemberTokenModel.getCode() == -1){
+            return validMemberTokenModel;
+        }
+        Member loginMember = memberService.findById(validMemberTokenModel.getData().getMemberId());
         if(loginMember.getId() != 1 && loginMember.getIsAdmin() == 1){
             return new ResponseModel(-1,"没有权限");
         }
@@ -83,13 +70,17 @@ public class MemberController extends BaseController {
      * @param id
      * @return
      */
-    @RequestMapping(value = "${managePath}/mem/manager/cancel/{id}",method = RequestMethod.GET)
+    @RequestMapping(value = "${managePath}/member/manager/cancel",method = RequestMethod.GET)
     @ResponseBody
-    public Object managerCancel(@PathVariable("id") Integer id) {
+    public Object managerCancel(@Param("id") Integer id) {
+        ResponseModel<MemberToken> validMemberTokenModel = validMemberToken();
+        if (validMemberTokenModel.getCode() == -1){
+            return validMemberTokenModel;
+        }
         if(id == null){
             return new ResponseModel(-1,"参数错误");
         }
-        Member loginMember = MemberUtil.getLoginMember(request);
+        Member loginMember = memberService.findById(validMemberTokenModel.getData().getMemberId());
         if(loginMember.getIsAdmin() == 1){
             return new ResponseModel(-1,"没有权限");
         }
@@ -97,45 +88,51 @@ public class MemberController extends BaseController {
     }
 
 
-    @RequestMapping("${managePath}/mem/index")
-    public String index(String key,Model model) {
+    @RequestMapping("${managePath}/member/list")
+    @ResponseBody
+    public Object list(String key) {
+        ResponseModel<MemberToken> validMemberTokenModel = validMemberToken();
+        if (validMemberTokenModel.getCode() == -1){
+            return validMemberTokenModel;
+        }
         Page page = new Page(request);
         ResponseModel responseModel = memberService.listByPage(page,key);
-        model.addAttribute("model",responseModel);
-        model.addAttribute("key",key);
-        return MANAGE_FTL_PATH + "index";
+        return responseModel;
     }
 
 
     /**
      * 会员启用、禁用操作
      */
-    @RequestMapping(value = "${managePath}/mem/member/isenable/{id}", method = RequestMethod.GET)
+    @RequestMapping(value = "${managePath}/member/isenable", method = RequestMethod.GET)
     @ResponseBody
-    public ResponseModel isenable(@PathVariable("id") int id) {
+    public ResponseModel isenable(@Param("id") int id) {
+        ResponseModel<MemberToken> validMemberTokenModel = validMemberToken();
+        if (validMemberTokenModel.getCode() == -1){
+            return validMemberTokenModel;
+        }
         return memberService.isenable(id);
     }
 
-    @RequestMapping(value = "${managePath}/mem/member/delete/{id}", method = RequestMethod.GET)
+    @RequestMapping(value = "${managePath}/member/delete", method = RequestMethod.GET)
     @ResponseBody
-    public Object delete(@PathVariable("id") int id) {
+    public Object delete(@Param("id") int id) {
+        ResponseModel<MemberToken> validMemberTokenModel = validMemberToken();
+        if (validMemberTokenModel.getCode() == -1){
+            return validMemberTokenModel;
+        }
         return memberService.delete(id);
     }
 
-    @RequestMapping(value = "${managePath}/mem/member/changepwd/{id}", method = RequestMethod.GET)
-    public String changepwd(@PathVariable("id") int id, Model model) {
-        Member member = memberService.findById(id);
-        if (member == null) {
-            return errorModel(model, "会员不存在");
-        }
-        model.addAttribute("member", member);
-        return MANAGE_FTL_PATH + "changepwd";
-    }
 
-    @RequestMapping(value = "${managePath}/mem/member/changepwd", method = RequestMethod.POST)
+    @RequestMapping(value = "${managePath}/member/changepwd", method = RequestMethod.POST)
     @ResponseBody
     public Object changepwd(int id, String password) {
-        Member loginMember = MemberUtil.getLoginMember(request);
+        ResponseModel<MemberToken> validMemberTokenModel = validMemberToken();
+        if (validMemberTokenModel.getCode() == -1){
+            return validMemberTokenModel;
+        }
+        Member loginMember = memberService.findById(validMemberTokenModel.getData().getMemberId());
         return memberService.changepwd(loginMember,id, password);
     }
